canjes devoluciones 98%

+++ app/src/main/java/union/union_vr1/AsyncTask/ExportMain.java
@@ -458,7 +458,7 @@ public class ExportMain extends AsyncTask<String, String, String> {
                             1,
                             cursorHistoVentaCreated.getInt(cursorHistoVentaCreated.getColumnIndexOrThrow(dbAdapter_histo_venta.HV_id_agente)),
                             2,
-                            cursorHistoVentaCreated.getInt(cursorHistoVentaCreated.getColumnIndexOrThrow(dbAdapter_histo_venta.HV_subtotal)),
+                            cursorHistoVentaCreated.getDouble(cursorHistoVentaCreated.getColumnIndexOrThrow(dbAdapter_histo_venta.HV_subtotal)),
                             cursorHistoVentaCreated.getString(cursorHistoVentaCreated.getColumnIndexOrThrow(dbAdapter_histo_venta.HV_id_histo))
                             );
